requirements.txt
fastapi
uvicorn
supabase
python-dotenv
requests
apscheduler